+++ blankiety_galantow/app.py
@@ -8,7 +8,6 @@ from fastapi import FastAPI, WebSocket
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from blankiety_galantow.routers import router
-from blankiety_galantow.connect_router import connect_router
 # Importing classes
 from .classes.Server import Server
 
@@ -21,9 +20,8 @@ app.mount("/js", StaticFiles(directory="resources/js"), name="js")
 app.mount("/vendor", StaticFiles(directory="resources/vendor"), name="vendor")
 app.mount("/game/css", StaticFiles(directory="resources/game/css"), name="game/css")
 app.mount("/game/js", StaticFiles(directory="resources/game/js"), name="game/js")
-# Attaching routers to app
+# Attaching router to app
 app.include_router(router)
-app.include_router(connect_router)
 
 # Starting server using uvicorn
 def run():
